% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_mapbox.R
\name{tm_mapbox}
\alias{tm_mapbox}
\title{Mapbox mode options}
\usage{
tm_mapbox(pitch, control.position, control.collapse, zoom)
}
\arguments{
\item{pitch}{The pitch angle}

\item{control.position}{The position of the layer control box}

\item{control.collapse}{Should the layer control box be collapsed?}

\item{zoom}{The zoom level of the map}
}
\value{
a [tmap::tmap-element]
}
\description{
Mapbox mode options. These options are specific to the mapbox mode.
}
\examples{
\dontrun{
library(tmap)
library(tmap.mapgl)
# getting API: https://walker-data.com/mapgl/articles/getting-started.html
# check API envir var: Sys.getenv("MAPBOX_PUBLIC_TOKEN")

tmap_mode("mapbox")
tm_shape(World) +
  tm_polygons("HPI", fill.scale = tm_scale_intervals(values = "brewer.rd_yl_gn"))

tm_shape(NLD_dist) +
  tm_polygons("employment_rate",
    fill.scale = tm_scale_intervals(values = "scico.roma"),
    lwd = 0.1) +
tm_shape(NLD_muni) +
  tm_polygons(fill = NULL, lwd = 1) +
tm_mapbox(pitch = 60) +
tm_basemap("mapbox.dark")
}
}
