% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_polygons_3d.R
\name{opt_tm_polygons3d}
\alias{opt_tm_polygons3d}
\title{Specify options to map layers}
\usage{
opt_tm_polygons3d(polygons.only = "ifany")
}
\arguments{
\item{polygons.only}{should only polygon geometries of the shape object (defined in [tm_shape()]) be plotted? By default `"ifany"`, which means `TRUE` in case a geometry collection is specified.}
}
\description{
The family of `opt_*()` functions can be used to specify options in the different `tm_*()`
functions.
}
\examples{
tm_shape(World) +
  tm_polygons(

   )

}
