% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_mapbox.R
\name{tm_mapbox}
\alias{tm_mapbox}
\title{Mapbox mode options}
\usage{
tm_mapbox(style, pitch)
}
\arguments{
\item{style}{To do: migrate with tmap styles and basemaps?}

\item{pitch}{The pitch angle}
}
\description{
Mapbox mode options. These options are specific to the mapbox mode.
}
\examples{
library(tmap)

tmap_mode("mapbox")
tm_shape(World) +
  tm_polygons("HPI", fill.scale = tm_scale_intervals(values = "brewer.rd_yl_gn"))

tm_shape(NLD_dist) +
	tm_polygons("employment_rate",
				fill.scale = tm_scale_intervals(values = "scico.roma"),
				lwd = 0.1) +
tm_shape(NLD_muni) +
	tm_polygons(fill = NULL, lwd = 1) +
	tm_deck(pitch = 75)
}
