% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgl_init.R, R/mapgl_shape.R, R/tmapMapbox.R,
%   R/tmapMapboxArrange.R, R/tmapMapboxAux.R, R/tmapMapboxComp.R,
%   R/tmapMapboxComp_charts.R, R/tmapMapboxLegend.R, R/tmapMapboxProviders.R,
%   R/tmapMapboxRun.R, R/tmapMapboxShiny.R, R/tmapMapboxXYlab.R,
%   R/tmapMapbox_layers.R, R/tmapMaplibreAux.R, R/tmapMaplibreComp.R,
%   R/tmapMaplibre_layers.R
\name{tmapMapboxInit}
\alias{tmapMapboxInit}
\alias{tmapMaplibreInit}
\alias{tmapMapboxAux}
\alias{tmapMaplibreAux}
\alias{tmapMapboxShape}
\alias{tmapMaplibreShape}
\alias{tmapMapboxOverlay}
\alias{tmapMaplibreOverlay}
\alias{tmapMapboxWrap}
\alias{tmapMapboxXtab}
\alias{tmapMaplibreWrap}
\alias{tmapMaplibreXtab}
\alias{tmapMapboxArrange}
\alias{tmapMapboxTilesPrep}
\alias{tmapMapboxTiles}
\alias{tmapMapboxGridPrep}
\alias{tmapMapboxGrid}
\alias{tmapMapboxGridXLab}
\alias{tmapMapboxGridYLab}
\alias{tmapMapboxCompPrepare.tm_title}
\alias{tmapMapboxCompHeight.tm_title}
\alias{tmapMapboxCompWidth.tm_title}
\alias{tmapMapboxLegPlot.tm_title}
\alias{tmapMapboxCompPrepare.tm_compass}
\alias{tmapMapboxCompHeight.tm_compass}
\alias{tmapMapboxCompWidth.tm_compass}
\alias{tmapMapboxLegPlot.tm_compass}
\alias{tmapMapboxCompPrepare.tm_scalebar}
\alias{tmapMapboxCompHeight.tm_scalebar}
\alias{tmapMapboxCompWidth.tm_scalebar}
\alias{tmapMapboxLegPlot.tm_scalebar}
\alias{tmapMapboxCompPrepare.tm_credits}
\alias{tmapMapboxCompHeight.tm_credits}
\alias{tmapMapboxCompWidth.tm_credits}
\alias{tmapMapboxLegPlot.tm_credits}
\alias{tmapMapboxCompPrepare.tm_mouse_coordinates}
\alias{tmapMapboxCompHeight.tm_mouse_coordinates}
\alias{tmapMapboxCompWidth.tm_mouse_coordinates}
\alias{tmapMapboxLegPlot.tm_mouse_coordinates}
\alias{tmapMapboxCompPrepare}
\alias{tmapMapboxCompHeight}
\alias{tmapMapboxCompWidth}
\alias{tmapMapboxLegPlot}
\alias{tmapMapboxLegPlot.tm_legend_standard_portrait}
\alias{tmapMapboxLegPlot.tm_legend_standard_landscape}
\alias{tmapMaplibreLegPlot.tm_legend_standard_portrait}
\alias{tmapMaplibreLegPlot.tm_legend_standard_landscape}
\alias{tmapMapboxComp}
\alias{tmapMapboxLegend}
\alias{tmapMaplibreComp}
\alias{tmapMapboxProviders}
\alias{tmapMaplibreProviders}
\alias{tmapMapboxRun}
\alias{tmapMaplibreRun}
\alias{renderTmapMapbox}
\alias{tmapOutputMapbox}
\alias{tmapProxyMapbox}
\alias{renderTmapMaplibre}
\alias{tmapOutputMaplibre}
\alias{tmapProxyMaplibre}
\alias{tmapMapboxXlab}
\alias{tmapMapboxYlab}
\alias{tmapMaplibreXlab}
\alias{tmapMaplibreYlab}
\alias{tmapMapboxPolygons}
\alias{tmapMaplibrePolygons}
\alias{tmapMapboxLines}
\alias{tmapMapboxSymbols}
\alias{tmapMapboxRaster}
\alias{tmapMapboxText}
\alias{tmapMaplibreTilesPrep}
\alias{tmapMaplibreTiles}
\alias{tmapMaplibreGridPrep}
\alias{tmapMaplibreGrid}
\alias{tmapMaplibreGridXLab}
\alias{tmapMaplibreGridYLab}
\alias{tmapMaplibreCompPrepare.tm_title}
\alias{tmapMaplibreCompHeight.tm_title}
\alias{tmapMaplibreCompWidth.tm_title}
\alias{tmapMaplibreLegPlot.tm_title}
\alias{tmapMaplibreCompPrepare.tm_compass}
\alias{tmapMaplibreCompHeight.tm_compass}
\alias{tmapMaplibreCompWidth.tm_compass}
\alias{tmapMaplibreLegPlot.tm_compass}
\alias{tmapMaplibreCompPrepare.tm_scalebar}
\alias{tmapMaplibreCompHeight.tm_scalebar}
\alias{tmapMaplibreCompWidth.tm_scalebar}
\alias{tmapMaplibreLegPlot.tm_scalebar}
\alias{tmapMaplibreCompPrepare.tm_credits}
\alias{tmapMaplibreCompHeight.tm_credits}
\alias{tmapMaplibreCompWidth.tm_credits}
\alias{tmapMaplibreLegPlot.tm_credits}
\alias{tmapMaplibreCompPrepare.tm_mouse_coordinates}
\alias{tmapMaplibreCompHeight.tm_mouse_coordinates}
\alias{tmapMaplibreCompWidth.tm_mouse_coordinates}
\alias{tmapMaplibreLegPlot.tm_mouse_coordinates}
\title{Internal tmap methods}
\usage{
tmapMapboxInit(o, return.asp = FALSE, vp, prx, dg = NULL, ...)

tmapMaplibreInit(o, return.asp = FALSE, vp, prx, dg = NULL, ...)

tmapMapboxAux(o, q)

tmapMaplibreAux(o, q)

tmapMapboxShape(bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreShape(bbx, facet_row, facet_col, facet_page, o)

tmapMapboxOverlay(bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreOverlay(bbx, facet_row, facet_col, facet_page, o)

tmapMapboxWrap(label, facet_row, facet_col, facet_page, o)

tmapMapboxXtab(label, facet_row, facet_col, facet_page, o)

tmapMaplibreWrap(label, facet_row, facet_col, facet_page, o)

tmapMaplibreXtab(label, facet_row, facet_col, facet_page, o)

tmapMapboxArrange(tms, nx, ncol, nrow, opts, knit, show, args, options)

tmapMapboxTilesPrep(a, bs, id, o)

tmapMapboxTiles(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapMapboxGridPrep(a, bs, id, o)

tmapMapboxGrid(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapMapboxGridXLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMapboxGridYLab(bi, bbx, facet_row, facet_col, facet_page, o)

\method{tmapMapboxCompPrepare}{tm_title}(comp, o)

\method{tmapMapboxCompHeight}{tm_title}(comp, o)

\method{tmapMapboxCompWidth}{tm_title}(comp, o)

\method{tmapMapboxLegPlot}{tm_title}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_compass}(comp, o)

\method{tmapMapboxCompHeight}{tm_compass}(comp, o)

\method{tmapMapboxCompWidth}{tm_compass}(comp, o)

\method{tmapMapboxLegPlot}{tm_compass}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_scalebar}(comp, o)

\method{tmapMapboxCompHeight}{tm_scalebar}(comp, o)

\method{tmapMapboxCompWidth}{tm_scalebar}(comp, o)

\method{tmapMapboxLegPlot}{tm_scalebar}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_credits}(comp, o)

\method{tmapMapboxCompHeight}{tm_credits}(comp, o)

\method{tmapMapboxCompWidth}{tm_credits}(comp, o)

\method{tmapMapboxLegPlot}{tm_credits}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_mouse_coordinates}(comp, o)

\method{tmapMapboxCompHeight}{tm_mouse_coordinates}(comp, o)

\method{tmapMapboxCompWidth}{tm_mouse_coordinates}(comp, o)

\method{tmapMapboxLegPlot}{tm_mouse_coordinates}(comp, m, o)

tmapMapboxCompPrepare(comp, o)

tmapMapboxCompHeight(comp, o)

tmapMapboxCompWidth(comp, o)

tmapMapboxLegPlot(comp, m, o)

\method{tmapMapboxLegPlot}{tm_legend_standard_portrait}(comp, m, o)

\method{tmapMapboxLegPlot}{tm_legend_standard_landscape}(comp, m, o)

tmapMaplibreLegPlot.tm_legend_standard_portrait(comp, m, o)

tmapMaplibreLegPlot.tm_legend_standard_landscape(comp, m, o)

tmapMapboxComp(
  comp,
  o,
  facet_row = NULL,
  facet_col = NULL,
  facet_page,
  class,
  stack,
  stack_auto,
  pos.h,
  pos.v,
  bbox
)

tmapMaplibreComp(
  comp,
  o,
  facet_row = NULL,
  facet_col = NULL,
  facet_page,
  class,
  stack,
  stack_auto,
  pos.h,
  pos.v,
  bbox
)

tmapMapboxProviders(credits)

tmapMaplibreProviders(credits)

tmapMapboxRun(o, q, show, knit, args)

tmapMaplibreRun(o, q, show, knit, args)

renderTmapMapbox(expr, env, quoted, execOnResize)

tmapOutputMapbox(outputId, width, height)

tmapProxyMapbox(mapId, session, x)

renderTmapMaplibre(expr, env, quoted, execOnResize)

tmapOutputMaplibre(outputId, width, height)

tmapProxyMaplibre(mapId, session, x)

tmapMapboxXlab(facet_page, o)

tmapMapboxYlab(facet_page, o)

tmapMaplibreXlab(facet_page, o)

tmapMaplibreYlab(facet_page, o)

tmapMapboxPolygons(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMaplibrePolygons(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxLines(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxSymbols(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxRaster(
  shpTM,
  dt,
  gp,
  pdt,
  popup.format,
  hdt,
  idt,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxText(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMaplibreTilesPrep(a, bs, id, o)

tmapMaplibreTiles(
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapMaplibreGridPrep(a, bs, id, o)

tmapMaplibreGrid(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapMaplibreGridXLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreGridYLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreCompPrepare.tm_title(comp, o)

tmapMaplibreCompHeight.tm_title(comp, o)

tmapMaplibreCompWidth.tm_title(comp, o)

tmapMaplibreLegPlot.tm_title(comp, m, o)

tmapMaplibreCompPrepare.tm_compass(comp, o)

tmapMaplibreCompHeight.tm_compass(comp, o)

tmapMaplibreCompWidth.tm_compass(comp, o)

tmapMaplibreLegPlot.tm_compass(comp, m, o)

tmapMaplibreCompPrepare.tm_scalebar(comp, o)

tmapMaplibreCompHeight.tm_scalebar(comp, o)

tmapMaplibreCompWidth.tm_scalebar(comp, o)

tmapMaplibreLegPlot.tm_scalebar(comp, m, o)

tmapMaplibreCompPrepare.tm_credits(comp, o)

tmapMaplibreCompHeight.tm_credits(comp, o)

tmapMaplibreCompWidth.tm_credits(comp, o)

tmapMaplibreLegPlot.tm_credits(comp, m, o)

tmapMaplibreCompPrepare.tm_mouse_coordinates(comp, o)

tmapMaplibreCompHeight.tm_mouse_coordinates(comp, o)

tmapMaplibreCompWidth.tm_mouse_coordinates(comp, o)

tmapMaplibreLegPlot.tm_mouse_coordinates(comp, m, o)

tmapMaplibrePolygons(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{o}{the list of options}

\item{return.asp}{return.asp}

\item{vp}{vp}

\item{q}{q}

\item{bbx}{bbx}

\item{facet_row, facet_col, facet_page}{row column and page id}

\item{label}{label}

\item{tms}{tmap objects}

\item{nx}{number of facets}

\item{ncol, }{nrow number of rows and columns}

\item{opts}{options}

\item{knit}{knit}

\item{show}{show}

\item{args}{args}

\item{options}{options}

\item{bs}{bs}

\item{id}{id}

\item{bi}{bi}

\item{pane}{pane}

\item{group}{group}

\item{comp}{the shape object}

\item{m}{mapbox oer maplibre object}

\item{class}{class}

\item{stack}{stack}

\item{stack_auto}{stack_auto}

\item{pos.h}{pos.h}

\item{pos.v}{pos.v}

\item{bbox}{bbox}

\item{credits}{credits}

\item{shpTM, }{dt, pdt, popup.format, hdt, idt, gp args}
}
\description{
Internal tmap methods
}
\keyword{internal}
