% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapMapbox.R, R/tmapMapboxArrange.R,
%   R/tmapMapboxAux.R, R/tmapMapboxComp_charts.R, R/tmapMapboxInit.R,
%   R/tmapMapboxLegend.R, R/tmapMapboxProviders.R, R/tmapMapboxRun.R,
%   R/tmapMapboxShape.R, R/tmapMapboxXYlab.R, R/tmapMapbox_layers.R
\name{tmapMapboxWrap}
\alias{tmapMapboxWrap}
\alias{tmapMapboxXtab}
\alias{tmapMapboxArrange}
\alias{tmapMapboxTilesPrep}
\alias{tmapMapboxTiles}
\alias{tmapMapboxGridPrep}
\alias{tmapMapboxGrid}
\alias{tmapMapboxGridXLab}
\alias{tmapMapboxGridYLab}
\alias{tmapMapboxCompPrepare}
\alias{tmapMapboxCompHeight}
\alias{tmapMapboxCompWidth}
\alias{tmapMapboxLegPlot}
\alias{tmapMapboxInit}
\alias{tmapMapboxAux}
\alias{tmapMapboxLegend}
\alias{tmapMapboxComp}
\alias{tmapMapboxProviders}
\alias{tmapMapboxRun}
\alias{tmapMapboxShape}
\alias{tmapMapboxOverlay}
\alias{tmapMapboxXlab}
\alias{tmapMapboxYlab}
\alias{tmapMapboxPolygons}
\alias{tmapMapboxLines}
\alias{tmapMapboxSymbols}
\alias{tmapMapboxRaster}
\alias{tmapMapboxText}
\title{Internal tmap methods}
\usage{
tmapMapboxWrap(label, facet_row, facet_col, facet_page, o)

tmapMapboxXtab(label, facet_row, facet_col, facet_page, o)

tmapMapboxArrange(tms, nx, ncol, nrow, opts, knit, show, args, options)

tmapMapboxTilesPrep(a, bs, id, o)

tmapMapboxTiles(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapMapboxGridPrep(a, bs, id, o)

tmapMapboxGrid(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapMapboxGridXLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMapboxGridYLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMapboxCompPrepare(comp, o)

tmapMapboxCompHeight(comp, o)

tmapMapboxCompWidth(comp, o)

tmapMapboxLegPlot(comp, mapbox, o)

tmapMapboxInit(o, return.asp = FALSE, vp, prx, dg = NULL, ...)

tmapMapboxAux(o, q)

tmapMapboxComp(
  comp,
  o,
  facet_row = NULL,
  facet_col = NULL,
  facet_page,
  class,
  stack,
  stack_auto,
  pos.h,
  pos.v,
  bbox
)

tmapMapboxProviders(credits)

tmapMapboxRun(o, q, show, knit, args)

tmapMapboxShape(bbx, facet_row, facet_col, facet_page, o)

tmapMapboxOverlay(bbx, facet_row, facet_col, facet_page, o)

tmapMapboxXlab(facet_page, o)

tmapMapboxYlab(facet_page, o)

tmapMapboxPolygons(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxLines(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapMapboxLines(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapMapboxSymbols(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxRaster(
  shpTM,
  dt,
  gp,
  pdt,
  popup.format,
  hdt,
  idt,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxText(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{label}{label}

\item{facet_row, facet_col, facet_page}{row column and page id}

\item{o}{the list of options}

\item{tms}{tmap objects}

\item{nx}{number of facets}

\item{ncol, }{nrow number of rows and columns}

\item{opts}{options}

\item{knit}{knit}

\item{show}{show}

\item{args}{args}

\item{options}{options}

\item{bs}{bs}

\item{id}{id}

\item{bi}{bi}

\item{bbx}{bbx}

\item{pane}{pane}

\item{group}{group}

\item{comp}{the shape object}

\item{mapbox}{mapbox object}

\item{return.asp}{return.asp}

\item{vp}{vp}

\item{q}{q}

\item{class}{class}

\item{stack}{stack}

\item{stack_auto}{stack_auto}

\item{pos.h}{pos.h}

\item{pos.v}{pos.v}

\item{bbox}{bbox}

\item{credits}{credits}

\item{shpTM, }{dt, pdt, popup.format, hdt, idt, gp args}
}
\description{
Internal tmap methods
}
\keyword{internal}
