% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variable_height.R, R/mapgl_arrange.R,
%   R/mapgl_aux.R, R/mapgl_comp.R, R/mapgl_comp_charts.R, R/mapgl_components.R,
%   R/mapgl_init.R, R/mapgl_legend.R, R/mapgl_providers.R, R/mapgl_run.R,
%   R/mapgl_shape.R, R/mapgl_shiny.R, R/mapgl_wrap_xtab.R, R/tm_aux_layers.R,
%   R/tmapMapboxDataPlot.R, R/tmapMaplibreDataPlot.R
\name{tmapValuesCheck_height}
\alias{tmapValuesCheck_height}
\alias{tmapValuesIsDiv_height}
\alias{tmapValuesRange_height}
\alias{tmapValuesVV_height}
\alias{tmapValuesSubmit_height}
\alias{tmapValuesScale_height}
\alias{tmapValuesColorize_height}
\alias{tmapValuesCVV_height}
\alias{tmapMapboxArrange}
\alias{tmapMaplibreArrange}
\alias{tmapMapboxAuxPrepare}
\alias{tmapMapboxAuxPrepare.default}
\alias{tmapMapboxAuxPlot}
\alias{tmapMapboxAuxPlot.default}
\alias{tmapMaplibreAuxPrepare}
\alias{tmapMaplibreAuxPrepare.default}
\alias{tmapMaplibreAuxPlot}
\alias{tmapMaplibreAuxPlot.default}
\alias{tmapMapboxAuxPrepare.tm_aux_basemap}
\alias{tmapMaplibreAuxPrepare.tm_aux_basemap}
\alias{tmapMapboxAuxPrepare.tm_aux_tiles}
\alias{tmapMaplibreAuxPrepare.tm_aux_tiles}
\alias{tmapMapboxAuxPlot.tm_aux_basemap}
\alias{tmapMaplibreAuxPlot.tm_aux_basemap}
\alias{tmapMapboxAuxPlot.tm_aux_tiles}
\alias{tmapMaplibreAuxPlot.tm_aux_tiles}
\alias{tmapMapboxGridXLab}
\alias{tmapMapboxGridYLab}
\alias{tmapMaplibreGridXLab}
\alias{tmapMaplibreGridYLab}
\alias{tmapMapboxCompPrepare.tm_title}
\alias{tmapMapboxCompHeight.tm_title}
\alias{tmapMapboxCompWidth.tm_title}
\alias{tmapMapboxCompPlot.tm_title}
\alias{tmapMapboxCompPrepare.tm_scalebar}
\alias{tmapMapboxCompHeight.tm_scalebar}
\alias{tmapMapboxCompWidth.tm_scalebar}
\alias{tmapMapboxCompPlot.tm_scalebar}
\alias{tmapMapboxCompPrepare.tm_snow}
\alias{tmapMapboxCompHeight.tm_snow}
\alias{tmapMapboxCompWidth.tm_snow}
\alias{tmapMapboxCompPlot.tm_snow}
\alias{tmapMaplibreCompPrepare.tm_title}
\alias{tmapMaplibreCompHeight.tm_title}
\alias{tmapMaplibreCompWidth.tm_title}
\alias{tmapMaplibreCompPlot.tm_title}
\alias{tmapMaplibreCompPrepare.tm_scalebar}
\alias{tmapMaplibreCompHeight.tm_scalebar}
\alias{tmapMaplibreCompWidth.tm_scalebar}
\alias{tmapMaplibreCompPlot.tm_scalebar}
\alias{tmapMaplibreCompPrepare.tm_credits}
\alias{tmapMaplibreCompHeight.tm_credits}
\alias{tmapMaplibreCompWidth.tm_credits}
\alias{tmapMaplibreCompPlot.tm_credits}
\alias{tmapMaplibreCompPrepare.tm_mouse_coordinates}
\alias{tmapMaplibreCompHeight.tm_mouse_coordinates}
\alias{tmapMaplibreCompWidth.tm_mouse_coordinates}
\alias{tmapMaplibreCompPlot.tm_mouse_coordinates}
\alias{tmapMapboxCompPrepare.tm_minimap}
\alias{tmapMaplibreCompPrepare.tm_minimap}
\alias{tmapMapboxCompHeight.tm_minimap}
\alias{tmapMapboxCompWidth.tm_minimap}
\alias{tmapMaplibreCompHeight.tm_minimap}
\alias{tmapMaplibreCompWidth.tm_minimap}
\alias{tmapMapboxCompPlot.tm_minimap}
\alias{tmapMaplibreCompPlot.tm_minimap}
\alias{tmapMapboxCompPrepare}
\alias{tmapMapboxCompHeight}
\alias{tmapMapboxCompWidth}
\alias{tmapMapboxCompPlot}
\alias{tmapMapboxCompPrepare.default}
\alias{tmapMapboxCompPrepare.tm_chart_none}
\alias{tmapMaplibreCompPrepare}
\alias{tmapMaplibreCompHeight}
\alias{tmapMaplibreCompWidth}
\alias{tmapMaplibreCompPlot}
\alias{tmapMaplibreCompPrepare.default}
\alias{tmapMaplibreCompPrepare.tm_chart_none}
\alias{tmapMapboxCompPrepare.tm_geocoder}
\alias{tmapMapboxCompHeight.tm_geocoder}
\alias{tmapMapboxCompWidth.tm_geocoder}
\alias{tmapMapboxCompPlot.tm_geocoder}
\alias{tmapMaplibreCompPrepare.tm_geocoder}
\alias{tmapMaplibreCompHeight.tm_geocoder}
\alias{tmapMaplibreCompWidth.tm_geocoder}
\alias{tmapMaplibreCompPlot.tm_geocoder}
\alias{tmapMapboxInit}
\alias{tmapMaplibreInit}
\alias{tmapMapboxAux}
\alias{tmapMaplibreAux}
\alias{tmapMapboxCompPlot.tm_legend_portrait}
\alias{tmapMapboxCompPlot.tm_legend_landscape}
\alias{tmapMaplibreCompPlot.tm_legend_portrait}
\alias{tmapMaplibreCompPlot.tm_legend_landscape}
\alias{tmapMapboxComp}
\alias{tmapMapboxLegend}
\alias{tmapMaplibreComp}
\alias{tmapMapboxProviders}
\alias{tmapMaplibreProviders}
\alias{tmapMapboxRun}
\alias{tmapMaplibreRun}
\alias{tmapMapboxShape}
\alias{tmapMaplibreShape}
\alias{tmapMapboxOverlay}
\alias{tmapMaplibreOverlay}
\alias{renderTmapMapbox}
\alias{tmapOutputMapbox}
\alias{tmapProxyMapbox}
\alias{renderTmapMaplibre}
\alias{tmapOutputMaplibre}
\alias{tmapProxyMaplibre}
\alias{tmapMapboxWrap}
\alias{tmapMapboxXtab}
\alias{tmapMaplibreWrap}
\alias{tmapMaplibreXtab}
\alias{tmapMapboxAuxPrepare.tm_aux_snow}
\alias{tmapMapboxAuxPlot.tm_aux_snow}
\alias{tmapMapboxAuxPrepare.tm_aux_rain}
\alias{tmapMapboxAuxPlot.tm_aux_rain}
\alias{tmapMapboxDataPlot}
\alias{tmapMapboxDataPlot.default}
\alias{tmapMapboxDataPlot.tm_data_polygons}
\alias{tmapMapboxDataPlot.tm_data_polygons_3d}
\alias{tmapMapboxDataPlot.tm_data_lines}
\alias{tmapMapboxDataPlot.tm_data_symbols}
\alias{tmapMapboxDataPlot.tm_data_raster}
\alias{tmapMapboxDataPlot.tm_data_fill}
\alias{tmapMapboxDataPlot.tm_data_borders}
\alias{tmapMapboxDataPlot.tm_data_dots}
\alias{tmapMapboxDataPlot.tm_data_bubbles}
\alias{tmapMaplibreDataPlot}
\alias{tmapMaplibreDataPlot.default}
\alias{tmapMaplibreDataPlot.tm_data_polygons}
\alias{tmapMaplibreDataPlot.tm_data_polygons_3d}
\alias{tmapMaplibreDataPlot.tm_data_lines}
\alias{tmapMaplibreDataPlot.tm_data_symbols}
\alias{tmapMaplibreDataPlot.tm_data_raster}
\alias{tmapMaplibreDataPlot.tm_data_fill}
\alias{tmapMaplibreDataPlot.tm_data_borders}
\alias{tmapMaplibreDataPlot.tm_data_dots}
\alias{tmapMaplibreDataPlot.tm_data_bubbles}
\title{Internal tmap methods}
\usage{
tmapValuesCheck_height(x, is_var = TRUE)

tmapValuesIsDiv_height(x)

tmapValuesRange_height(x, n, isdiv)

tmapValuesVV_height(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesSubmit_height(x, args)

tmapValuesScale_height(x, scale)

tmapValuesColorize_height(x, pc)

tmapValuesCVV_height(x, value.na, n, range, scale, rep, o)

tmapMapboxArrange(tms, nx, ncol, nrow, opts, knit, show, args, options)

tmapMaplibreArrange(tms, nx, ncol, nrow, opts, knit, show, args, options)

tmapMapboxAuxPrepare(a, bs, id, o)

\method{tmapMapboxAuxPrepare}{default}(a, bs, id, o)

tmapMapboxAuxPlot(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMapboxAuxPlot}{default}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapMaplibreAuxPrepare(a, bs, id, o)

\method{tmapMaplibreAuxPrepare}{default}(a, bs, id, o)

tmapMaplibreAuxPlot(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMaplibreAuxPlot}{default}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMapboxAuxPrepare}{tm_aux_basemap}(a, bs, id, o)

\method{tmapMaplibreAuxPrepare}{tm_aux_basemap}(a, bs, id, o)

\method{tmapMapboxAuxPrepare}{tm_aux_tiles}(a, bs, id, o)

\method{tmapMaplibreAuxPrepare}{tm_aux_tiles}(a, bs, id, o)

\method{tmapMapboxAuxPlot}{tm_aux_basemap}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMaplibreAuxPlot}{tm_aux_basemap}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMapboxAuxPlot}{tm_aux_tiles}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMaplibreAuxPlot}{tm_aux_tiles}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapMapboxGridXLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMapboxGridYLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreGridXLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreGridYLab(bi, bbx, facet_row, facet_col, facet_page, o)

\method{tmapMapboxCompPrepare}{tm_title}(comp, o)

\method{tmapMapboxCompHeight}{tm_title}(comp, o)

\method{tmapMapboxCompWidth}{tm_title}(comp, o)

\method{tmapMapboxCompPlot}{tm_title}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_scalebar}(comp, o)

\method{tmapMapboxCompHeight}{tm_scalebar}(comp, o)

\method{tmapMapboxCompWidth}{tm_scalebar}(comp, o)

\method{tmapMapboxCompPlot}{tm_scalebar}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_snow}(comp, o)

\method{tmapMapboxCompHeight}{tm_snow}(comp, o)

\method{tmapMapboxCompWidth}{tm_snow}(comp, o)

\method{tmapMapboxCompPlot}{tm_snow}(comp, m, o)

\method{tmapMaplibreCompPrepare}{tm_title}(comp, o)

\method{tmapMaplibreCompHeight}{tm_title}(comp, o)

\method{tmapMaplibreCompWidth}{tm_title}(comp, o)

\method{tmapMaplibreCompPlot}{tm_title}(comp, m, o)

\method{tmapMaplibreCompPrepare}{tm_scalebar}(comp, o)

\method{tmapMaplibreCompHeight}{tm_scalebar}(comp, o)

\method{tmapMaplibreCompWidth}{tm_scalebar}(comp, o)

\method{tmapMaplibreCompPlot}{tm_scalebar}(comp, m, o)

\method{tmapMaplibreCompPrepare}{tm_credits}(comp, o)

\method{tmapMaplibreCompHeight}{tm_credits}(comp, o)

\method{tmapMaplibreCompWidth}{tm_credits}(comp, o)

\method{tmapMaplibreCompPlot}{tm_credits}(comp, m, o)

\method{tmapMaplibreCompPrepare}{tm_mouse_coordinates}(comp, o)

\method{tmapMaplibreCompHeight}{tm_mouse_coordinates}(comp, o)

\method{tmapMaplibreCompWidth}{tm_mouse_coordinates}(comp, o)

\method{tmapMaplibreCompPlot}{tm_mouse_coordinates}(comp, m, o)

\method{tmapMapboxCompPrepare}{tm_minimap}(comp, o)

\method{tmapMaplibreCompPrepare}{tm_minimap}(comp, o)

\method{tmapMapboxCompHeight}{tm_minimap}(comp, o)

\method{tmapMapboxCompWidth}{tm_minimap}(comp, o)

\method{tmapMaplibreCompHeight}{tm_minimap}(comp, o)

\method{tmapMaplibreCompWidth}{tm_minimap}(comp, o)

\method{tmapMapboxCompPlot}{tm_minimap}(comp, m, o)

\method{tmapMaplibreCompPlot}{tm_minimap}(comp, m, o)

tmapMapboxCompPrepare(comp, o)

tmapMapboxCompHeight(comp, o)

tmapMapboxCompWidth(comp, o)

tmapMapboxCompPlot(comp, m, o)

\method{tmapMapboxCompPrepare}{default}(comp, o)

\method{tmapMapboxCompPrepare}{tm_chart_none}(comp, o)

tmapMaplibreCompPrepare(comp, o)

tmapMaplibreCompHeight(comp, o)

tmapMaplibreCompWidth(comp, o)

tmapMaplibreCompPlot(comp, m, o)

\method{tmapMaplibreCompPrepare}{default}(comp, o)

\method{tmapMaplibreCompPrepare}{tm_chart_none}(comp, o)

\method{tmapMapboxCompPrepare}{tm_geocoder}(comp, o)

\method{tmapMapboxCompHeight}{tm_geocoder}(comp, o)

\method{tmapMapboxCompWidth}{tm_geocoder}(comp, o)

\method{tmapMapboxCompPlot}{tm_geocoder}(comp, m, o)

\method{tmapMaplibreCompPrepare}{tm_geocoder}(comp, o)

\method{tmapMaplibreCompHeight}{tm_geocoder}(comp, o)

\method{tmapMaplibreCompWidth}{tm_geocoder}(comp, o)

\method{tmapMaplibreCompPlot}{tm_geocoder}(comp, m, o)

tmapMapboxInit(o, return.asp = FALSE, vp, prx, dg = NULL, ...)

tmapMaplibreInit(o, return.asp = FALSE, vp, prx, dg = NULL, ...)

tmapMapboxAux(o, q)

tmapMaplibreAux(o, q)

\method{tmapMapboxCompPlot}{tm_legend_portrait}(comp, m, o)

\method{tmapMapboxCompPlot}{tm_legend_landscape}(comp, m, o)

\method{tmapMaplibreCompPlot}{tm_legend_portrait}(comp, m, o)

\method{tmapMaplibreCompPlot}{tm_legend_landscape}(comp, m, o)

tmapMapboxComp(
  comp,
  o,
  facet_row = NULL,
  facet_col = NULL,
  facet_page,
  class,
  stack,
  stack_auto,
  pos.h,
  pos.v,
  bbox
)

tmapMaplibreComp(
  comp,
  o,
  facet_row = NULL,
  facet_col = NULL,
  facet_page,
  class,
  stack,
  stack_auto,
  pos.h,
  pos.v,
  bbox
)

tmapMapboxProviders(credits)

tmapMaplibreProviders(credits)

tmapMapboxRun(o, q, show, knit, knit_opts, args)

tmapMaplibreRun(o, q, show, knit, knit_opts, args)

tmapMapboxShape(bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreShape(bbx, facet_row, facet_col, facet_page, o)

tmapMapboxOverlay(bbx, facet_row, facet_col, facet_page, o)

tmapMaplibreOverlay(bbx, facet_row, facet_col, facet_page, o)

renderTmapMapbox(expr, env, quoted, execOnResize)

tmapOutputMapbox(outputId, width, height)

tmapProxyMapbox(mapId, session, x)

renderTmapMaplibre(expr, env, quoted, execOnResize)

tmapOutputMaplibre(outputId, width, height)

tmapProxyMaplibre(mapId, session, x)

tmapMapboxWrap(label, facet_row, facet_col, facet_page, o)

tmapMapboxXtab(label, facet_row, facet_col, facet_page, o)

tmapMaplibreWrap(label, facet_row, facet_col, facet_page, o)

tmapMaplibreXtab(label, facet_row, facet_col, facet_page, o)

\method{tmapMapboxAuxPrepare}{tm_aux_snow}(a, bs, id, o)

\method{tmapMapboxAuxPlot}{tm_aux_snow}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapMapboxAuxPrepare}{tm_aux_rain}(a, bs, id, o)

\method{tmapMapboxAuxPlot}{tm_aux_rain}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapMapboxDataPlot(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{default}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_polygons}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_polygons_3d}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_lines}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_symbols}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_raster}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_fill}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_borders}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_dots}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMapboxDataPlot}{tm_data_bubbles}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMaplibreDataPlot(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{default}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_polygons}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_polygons_3d}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_lines}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_symbols}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_raster}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_fill}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_borders}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_dots}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapMaplibreDataPlot}{tm_data_bubbles}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{o}{the list of options}

\item{args}{args}

\item{tms}{tmap objects}

\item{nx}{number of facets}

\item{ncol, }{nrow number of rows and columns}

\item{opts}{options}

\item{knit}{knit}

\item{show}{show}

\item{options}{options}

\item{bs}{bs}

\item{id}{id}

\item{bi}{bi}

\item{bbx}{bbx}

\item{facet_row, facet_col, facet_page}{row column and page id}

\item{pane}{pane}

\item{group}{group}

\item{comp}{the shape object}

\item{m}{mapbox oer maplibre object}

\item{return.asp}{return.asp}

\item{vp}{vp}

\item{q}{q}

\item{class}{class}

\item{stack}{stack}

\item{stack_auto}{stack_auto}

\item{pos.h}{pos.h}

\item{pos.v}{pos.v}

\item{bbox}{bbox}

\item{credits}{credits}

\item{knit_opts}{knit options}

\item{label}{label}

\item{shpTM, }{dt, pdt, popup.format, hdt, idt, gp args}
}
\value{
internal tmap lists
}
\description{
Internal tmap methods
}
\keyword{internal}
